requirements.txt
streamlit==1.17.0
spacy==3.2.4
pdfminer.six==20201018
nltk==3.7
openai==0.27.0
sentence-transformers==2.2.0
wordcloud==1.8.2.2
matplotlib==3.4.3
requests==2.26.0
